+++ db4e/Panes/XMRigPane.py
@@ -22,7 +22,7 @@ from db4e.Constants.Buttons import(
     ENABLE_LABEL, DISABLE_BUTTON_FIELD, DISABLE_LABEL,
     UPDATE_BUTTON_FIELD, NEW_BUTTON_FIELD, NEW_LABEL, UPDATE_LABEL)
 from db4e.Constants.Fields import (
-    ADD_DEPLOYMENT_FIELD, NEW_FIELD,
+    ADD_DEPLOYMENT_FIELD, NEW_FIELD, DEPLOYMENT_MGR_FIELD,
     FORM_3_FIELD, FORM_INPUT_15_FIELD, DISABLE_FIELD, FORM_INTRO_FIELD,
     FORM_LABEL_FIELD, HEALTH_BOX_FIELD, ELEMENT_FIELD,
     INSTANCE_FIELD, ENABLE_FIELD, OPS_MGR_FIELD,
@@ -155,50 +155,46 @@ class XMRigPane(Container):
 
         if button_id == NEW_BUTTON_FIELD:
             form_data = {
-                ELEMENT_TYPE_FIELD: XMRIG_FIELD,
                 TO_MODULE_FIELD: OPS_MGR_FIELD,
                 TO_METHOD_FIELD: ADD_DEPLOYMENT_FIELD,
+                ELEMENT_TYPE_FIELD: XMRIG_FIELD,
                 ELEMENT_FIELD: self.xmrig
             }
 
         elif button_id == UPDATE_BUTTON_FIELD:
             form_data = {
-                ELEMENT_TYPE_FIELD: XMRIG_FIELD,
-                TO_MODULE_FIELD: JOB_QUEUE_FIELD,
+                TO_MODULE_FIELD: DEPLOYMENT_MGR_FIELD,
                 TO_METHOD_FIELD: POST_JOB_FIELD,
                 OP_FIELD: UPDATE_FIELD,
+                ELEMENT_TYPE_FIELD: XMRIG_FIELD,
                 ELEMENT_FIELD: self.xmrig,
-                INSTANCE_FIELD: self.xmrig.instance()
             }
 
         elif button_id == ENABLE_BUTTON_FIELD:
             form_data = {
-                ELEMENT_TYPE_FIELD: XMRIG_FIELD,
-                TO_MODULE_FIELD: JOB_QUEUE_FIELD,
+                TO_MODULE_FIELD: DEPLOYMENT_MGR_FIELD,
                 TO_METHOD_FIELD: POST_JOB_FIELD,
                 OP_FIELD: ENABLE_FIELD,
+                ELEMENT_TYPE_FIELD: XMRIG_FIELD,
                 ELEMENT_FIELD: self.xmrig,
-                INSTANCE_FIELD: self.xmrig.instance()
             }
 
         elif button_id == DISABLE_BUTTON_FIELD:
             form_data = {
-                ELEMENT_TYPE_FIELD: XMRIG_FIELD,
-                TO_MODULE_FIELD: JOB_QUEUE_FIELD,
+                TO_MODULE_FIELD: DEPLOYMENT_MGR_FIELD,
                 TO_METHOD_FIELD: POST_JOB_FIELD,
                 OP_FIELD: DISABLE_FIELD,
+                ELEMENT_TYPE_FIELD: XMRIG_FIELD,
                 ELEMENT_FIELD: self.xmrig,
-                INSTANCE_FIELD: self.xmrig.instance()
             }
 
         elif button_id == DELETE_BUTTON_FIELD:
             form_data = {
-                ELEMENT_TYPE_FIELD: XMRIG_FIELD,
-                TO_MODULE_FIELD: JOB_QUEUE_FIELD,
+                TO_MODULE_FIELD: DEPLOYMENT_MGR_FIELD,
                 TO_METHOD_FIELD: POST_JOB_FIELD,
                 OP_FIELD: DELETE_FIELD,
+                ELEMENT_TYPE_FIELD: XMRIG_FIELD,
                 ELEMENT_FIELD: self.xmrig,
-                INSTANCE_FIELD: self.xmrig.instance()
             }            
 
         self.app.post_message(Db4eMsg(self, form_data=form_data))